requirements.txt
GitPython>=2.1.3
py-postgresql>=1.2.1
aiohttp>=2.0.7
